% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XappDEF.R
\name{GUI}
\alias{GUI}
\title{Generate GUI interface}
\usage{
GUI()
}
\value{
shiny app
}
\description{
GUI function

This app allows to generate a GUI that plots the data of the Temperature, CO2 and Milankovich cyles
as taken from EPICA DOME C
}
\examples{
\dontrun{
GUI()
}
}
