% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XappDEF.R
\name{sGUI}
\alias{sGUI}
\title{Generate GUI interface}
\usage{
sGUI()
}
\value{
shiny app
}
\description{
GUI function

This app allows to generate a GUI that plots the data of the Temperature, CO2 and Milankovich cyles
as taken from EPICA DOME C
}
\examples{
\dontrun{
sGUI()
}
}
